+++ src/lib/request-handlers/regions/put.ts
@@ -2,8 +2,8 @@ import { getRepository } from 'typeorm';
 import { Region } from '../../../orm/entity/Region';
 import { SUCCESS } from '../../messages';
 import { HttpCodes,  putResponse } from '../../types-interfaces';
+import { createMergeObj } from '../../utils/regions-helper';
 import { errorResponse, responder, responderWrongId, successResponse } from '../responders';
-import { createMergeObj } from './regions-helper';
 
 export const putRegion: putResponse = async (request, response) => {
   try {
